removed LambdaAuthorizerHandler, functionalities moved to AuthorizerPackage
